backend/charter/pyproject.toml
[project]
name = "charter"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "career-database",
    "boto3>=1.40.9",
    "langfuse>=3.3.4",
    "openai-agents[litellm]>=0.2.6",

    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "tenacity>=9.1.2",
    "httpx<0.28.0",
]

[tool.uv.sources]
career-database = { path = "../database", editable = true }
